Renamed searchPaxList to searchPassengersList.
